Unify Test Helper code into YorkieTestHelper test target
