+++ Tests/Integration/TreeIntegrationTests.swift
@@ -16,6 +16,9 @@
 
 import XCTest
 @testable import Yorkie
+#if SWIFT_TEST
+@testable import YorkieTestHelper
+#endif
 
 /**
  * `createChangePack` is a helper function that creates a change pack from the
